Started adding Windhaw's store stuff

+++ Assets/Player.cs
@@ -19,7 +19,7 @@ public class Player : Photon.Pun.MonoBehaviourPun
     public State status;
     public int statusTimer;
     public int health;
-    public decimal rubies = 0.00m;
+    public int rubies = 0;
     public Weapon currentWeapon;
     public List<Weapon> inventory;
     //public string armor;
